Move HealthIcon to common
