+++ assets/js/common/HealthIcon/HealthIcon.jsx
