+++ assets/js/common/HealthIcon/HealthIcon.stories.jsx
