+++ assets/js/common/HealthIcon/index.js
