+++ assets/js/pages/ClusterDetails/ExecutionIcon.jsx
@@ -3,7 +3,7 @@ import { computedIconCssClass } from '@lib/icon';
 
 import { EOS_SCHEDULE } from 'eos-icons-react';
 
-import HealthIcon from '@pages/HealthIcon';
+import HealthIcon from '@common/HealthIcon';
 import Spinner from '@common/Spinner';
 
 export function ExecutionIcon({ health, executionState, centered = false }) {
