+++ assets/js/pages/ExecutionResults/CheckResultDetail/CheckDetailHeader.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import HealthIcon from '@pages/HealthIcon';
+import HealthIcon from '@common/HealthIcon';
 import { clusterWarningBanner } from '@pages/ExecutionResults/ExecutionHeader';
 import CheckResultInfoBox from './CheckResultInfoBox';
 import { isTargetCluster } from '../checksUtils';
