+++ assets/js/pages/ExecutionResults/ExecutionResults.jsx
@@ -3,7 +3,7 @@ import ReactMarkdown from 'react-markdown';
 import remarkGfm from 'remark-gfm';
 import { getHostID } from '@state/selectors/host';
 import Accordion from '@common/Accordion';
-import HealthIcon from '@pages/HealthIcon';
+import HealthIcon from '@common/HealthIcon';
 import Modal from '@common/Modal';
 import PremiumPill from '@common/PremiumPill';
 import Table from '@common/Table';
