+++ assets/js/pages/HostDetailsPage/HostSummary.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { chunk } from 'lodash';
 
 import ClusterLink from '@pages/ClusterDetails/ClusterLink';
-import HealthIcon from '@pages/HealthIcon';
+import HealthIcon from '@common/HealthIcon';
 import ListView from '@common/ListView';
 import Tooltip from '@common/Tooltip';
